requirements.txt
aiohttp==1.0.5
astroid==1.5.3
async-timeout==2.0.0
chardet==3.0.4
colorama==0.3.9
discord.py==0.16.12
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==3.3.2
pylint==1.7.4
rope==0.10.7
six==1.11.0
websockets==3.4
wrapt==1.10.11
